Split up protocol.cpp
protocol.cpp becomes too big and additional tests for SUL and index
records will be added soon. Hence splitting tests into several files.
