+++ lib/test/sul.cpp
@@ -1,0 +1,204 @@
+#include <cstring>
+#include <sstream>
+#include <string>
+#include <cstdint>
+
+#include <catch2/catch.hpp>
+
+#include <dlisio/dlisio.h>
+
+
+struct SULv1 {
+    int seqnum          = -1;
+    int major           = -1;
+    int minor           = -1;
+    int layout          = -1;
+    std::int64_t maxlen = -1;
+    char id[ 61 ] = {};
+
+    bool operator==( const SULv1& o ) const {
+        std::string sid = this->id;
+        return this->seqnum == o.seqnum
+            && this->major  == o.major
+            && this->minor  == o.minor
+            && this->layout == o.layout
+            && this->maxlen == o.maxlen
+            &&       sid    == o.id
+            ;
+    }
+
+    bool operator!=( const SULv1& o ) const {
+        return !(*this == o);
+    }
+};
+
+std::ostream& operator<<( std::ostream& stream, const SULv1& x ) {
+    return stream << "SULv1 {"
+                  << "\n\t.seq = " << x.seqnum
+                  << "\n\t.ver = " << x.major << "." << x.minor
+                  << "\n\t.lay = " << (x.layout == DLIS_STRUCTURE_RECORD ?
+                                       "RECORD" : "UNKNOWN")
+                  << "\n\t.len = " << x.maxlen
+                  << "\n\t.idt = '" << x.id << "'"
+                  << "\n}";
+
+}
+
+std::string blockify( const std::string& xs, int len = 16 ) {
+    /*
+     * Print the string to disk, formatted like hexdump -c would, except
+     * embedded NUL bytes are printed as @
+     * Unfortunately, this makes intational @'s indistinguishable from \0, but
+     * it should be good enough for now
+     */
+    int pos = 0;
+    std::stringstream ss;
+
+    while( true ) {
+        if( pos + len > xs.size() ) {
+            ss.write( xs.c_str() + pos, xs.size() - pos );
+            ss << "\n";
+            return ss.str();
+        }
+
+        if( pos + len == xs.size() ) {
+            return ss.str();
+        }
+
+        std::string line( xs, pos, len );
+        for( auto& c : line ) if( c == '\0' ) c = '@';
+        ss << "|" << line << "|\n";
+        pos += len;
+    }
+}
+
+SULv1 parse_sulv1( const std::string& sul ) {
+    SULv1 v1;
+    auto err = dlis_sul( sul.c_str(), &v1.seqnum,
+                                      &v1.major,
+                                      &v1.minor,
+                                      &v1.layout,
+                                      &v1.maxlen,
+                                      v1.id );
+    REQUIRE( err == 0 );
+    return v1;
+}
+
+TEST_CASE("A simple, well-formatted SULv1", "[sul][v1]") {
+
+    SULv1 ref;
+    ref.seqnum = 1;
+    ref.major  = 1;
+    ref.minor  = 0;
+    ref.layout = DLIS_STRUCTURE_RECORD;
+    ref.maxlen = 8192;
+    std::strcpy( ref.id, "Default Storage Set "
+                         "                    "
+                         "                    " );
+
+    /*
+     * Generate all valid permutations of numbers with paddings and leading
+     * zeros and whatnot, that would produce the same record, and create a test
+     * case for each of them. Since revision and record only has one valid
+     * format (for revision 1.0), they don't change. The storage set identifier
+     * is an arbitrary string anyway so there's no wiggle room for two distinct
+     * strings that produce the same record
+     */
+
+    static const std::string sequence_numbers[] = {
+        "   1",
+        "  01",
+        " 001",
+        "0001",
+    };
+
+    static const std::string revisions[] = {
+        "V1.00",
+    };
+
+    static const std::string structures[] = {
+        "RECORD",
+    };
+
+    static const std::string maxlens[] = {
+        " 8192",
+        "08192",
+    };
+
+    static const std::string identifiers[] = {
+        "Default Storage Set "
+        "                    "
+        "                    ",
+    };
+
+    for( const auto& seq : sequence_numbers )
+    for( const auto& rev : revisions )
+    for( const auto& rec : structures )
+    for( const auto& len : maxlens )
+    for( const auto& idt : identifiers )
+    SECTION("[" + seq + "," + len + "]" ) {
+        const auto label = seq + rev + rec + len + idt;
+        INFO( "Storage unit label:\n" << blockify( label ) );
+
+        auto v1 = parse_sulv1( label );
+        CHECK( v1 == ref );
+    }
+}
+
+TEST_CASE("A well-formatted SULv1 with undefined maxlen", "[sul][v1]") {
+    SULv1 ref;
+    ref.seqnum = 1;
+    ref.major  = 1;
+    ref.minor  = 0;
+    ref.layout = DLIS_STRUCTURE_RECORD;
+    ref.maxlen = 0;
+    std::strcpy( ref.id, "Default Storage Set "
+                         "                    "
+                         "                    " );
+
+    static const std::string pre = "   1"
+                                   "V1.00"
+                                   "RECORD";
+
+    static const std::string post = "Default Storage Set "
+                                    "                    "
+                                    "                    ";
+
+    static const std::string maxlens[] = {
+        "    0",
+        "   00",
+        "  000",
+        " 0000",
+        "00000",
+    };
+
+    for( const auto& len : maxlens )
+    SECTION("Explicit zero: '" + len + "'") {
+        const auto label = pre + len + post;
+        INFO( "Storage unit label:\n" << blockify( label ) );
+
+        auto v1 = parse_sulv1( label );
+        CHECK( v1 == ref );
+    }
+
+    /*
+     * embedded NULL-bytes aren't according to spec, but they're trivially
+     * handled by the same code, so check their behaviour
+     */
+    static const std::string maxlens0[] = {
+        { "0\0000", 5 },
+        { "00\000", 5 },
+        { "000\00", 5 },
+        { " 00\0 ", 5 }, // trailing spaces after early \0 is ok
+        { "  0\0 ", 5 },
+    };
+
+    for( const auto& len : maxlens0 )
+    SECTION("Zeros with embedded NUL: '" + std::string( len.c_str() ) + "'") {
+        const auto label = pre + len + post;
+        INFO( "Storage unit label:\n" << blockify( label ) );
+
+        auto v1 = parse_sulv1( label );
+        CHECK( v1 == ref );
+    }
+}
